INFUND-7367 - Addresses PR comments
Former-commit-id: 099b192f515e59cee18a30a4658b1be1965a38a3

+++ ifs-data-service/src/main/java/org/innovateuk/ifs/competition/transactional/CompetitionKeyStatisticsServiceImpl.java
@@ -73,7 +73,7 @@ public class CompetitionKeyStatisticsServiceImpl extends BaseTransactionalServic
                 .filter(as -> as.getAssessors() < competitionKeyStatisticsResource.getApplicationsPerAssessor())
                 .mapToInt(e -> 1)
                 .sum());
-        competitionKeyStatisticsResource.setAssessorsWithoutApplications((int) competitionParticipantRepository.getByCompetitionIdAndRoleAndStatus(competitionId, ASSESSOR, ParticipantStatus.ACCEPTED)
+        competitionKeyStatisticsResource.setAssessorsWithoutApplications(competitionParticipantRepository.getByCompetitionIdAndRoleAndStatus(competitionId, ASSESSOR, ParticipantStatus.ACCEPTED)
                 .stream()
                 .filter(cp -> assessmentRepository.countByParticipantUserIdAndActivityStateStateNotIn(cp.getId(), of(REJECTED, WITHDRAWN)) == 0)
                 .mapToInt(e -> 1)
